+++ apps/carp_mobile_sensing_app/lib/src/sensing/sensing.dart
@@ -114,17 +114,18 @@ class Sensing {
 
     await client!.configure();
 
-    // add and deploy this deployment
+    // add and deploy this study deployment
     _controller =
         await client!.addStudy(bloc.studyDeploymentId!, deviceRolename!);
+    await _controller?.tryDeployment();
 
     // configure the controller
-    await _controller!.configure();
+    await _controller?.configure();
 
     // controller.resume();
 
     // listening on the data stream and print them as json to the debug console
-    _controller!.data.listen((data) => print(toJsonString(data)));
+    _controller?.data.listen((data) => print(toJsonString(data)));
 
     info('$runtimeType initialized');
   }